+++ harness/determined/cli/__init__.py
@@ -13,6 +13,7 @@ from determined.cli import (
     agent,
     checkpoint,
     cli,
+    ntsc,
     command,
     experiment,
     master,
@@ -20,7 +21,6 @@ from determined.cli import (
     notebook,
     project,
     rbac,
-    remote,
     render,
     resources,
     shell,
